Private API

Internals

Database

GitHubAPI.db_host — Constant.

db_host = "sdad.policy-analytics.net"

Host for the database.

source

GitHubAPI.db_port — Constant.

db_port = 5434

Port for the postgis_1 in the database.

source

GitHubAPI.dbname — Constant.

dbname = "oss"

Database for the Open-Source Software project.

source

GitHubAPI.db_pwd — Constant.

db_pwd::String

The password for the database.

source

GitHubAPI.db_user — Constant.

db_user::String

The username for the database (i.e., your UVA computing ID).

source

GitHubAPI.postgis_dtf — Constant.

postgis_dtf = dateformat"yyyy-mm-dd HH:MM:SSzzzz"

HTTP responses require this datetime format.

source

GitHubAPI.dbconnect — Function.

dbconnect()::Connection

Returns a connection to the postgis_1 OSS database (i.e., your UVA computing ID).

Example

conn = dbconnect()

source

GitHubAPI.insert_record_repos_by_license! — Function.

insert_record_repos_by_license!(conn::Connection,
                                license::AbstractString,
                                created_at::AbstractString,
                                as_of::AbstractString,
                                node)

It inserts each record to the universe.github_repos table.

source

GitHubAPI.insert_commit! — Function.

insert_commit!(conn::Connection, slug::AbstractString, node)

It inserts each record to the universe.github_commits table.

source

GitHub API

GitHubAPI.github_endpoint — Constant.

github_endpoint = "https://api.github.com/graphql"

Endpoint for the GitHub API v4 (GraphQL).

source

GitHubAPI.github_login — Constant.

github_login::String

Your GitHub handle.

source

GitHubAPI.github_header — Constant.

github_header = Dict("User-Agent" => github_login)

Header for the GitHub API.

source

GitHubAPI.client — Constant.

client = GraphQLClient(github_endpoint,
                       auth = "bearer $github_token",
                       headers = github_header)

Client for the GitHub API.

source

GitHubAPI.github_api_query — Constant.

github_api_query::String

Queries for finding open-sourced projects and their commit information from GitHub.

source

GitHubAPI.until — Constant.

until::ZonedDateTime

Until when should the scrapper query data. Currently at "2019-08-15T00:00:00-04:00".

source

GitHubAPI.binary_search_dt_interval — Function.

binary_search_dt_interval(license::AbstractString,
                          interval::AbstractString)::data, as_of, created_at

Given a license and a datetime interval, it will use binary search to find a datetime interval with no more than 1,000 results.

source

GitHubAPI.github_wait_out — Function.

github_wait_out(rateLimit)

If the GitHub personal access token has exhausted the per hour limit, it waits until it resets.

source

GitHubAPI.response_dtf — Constant.

response_dtf = dateformat"d u y H:M:S Z"

HTTP responses require this datetime format.

source

GitHubAPI.github_dtf — Constant.

github_dtf = "yyyy-mm-ddTHH:MM:SSzzzz"

GitHub zoned datetime format.

source

GitHubAPI.get_as_of — Function.

get_as_of(response::Response)::String

Returns the zoned date time when the response was returned.

source