Public API

Public API

Access the application

]dev https://github.com/uva-bi-sdad/GitHubAPI.jl

GitHubAPI — Module.

GitHubAPI

Module used by the Social Decision and Analytics Division (SDAD) of the Biocomplexity Institute and Initiative of the University of Virginia.

This module was designed for the Open-Source Software (OSS) project.

source

Set up your config file (confs/config.simple)

GitHubAPI.sdad_setup! — Function.

sdad_setup!(;db_user::AbstractString = "",
             db_pwd::AbstractString = "",
             github_login::AbstractString = "",
             github_token::AbstractString = "")

This function writes the configurations to confs/config.simple

Arguments

db_user: Your username for the postgis_1 database (i.e., your UVA computing ID).

db_pwd: The password for the postgis_1 database.

github_login: Your GitHub login (handle).

github_token: A 40 alphanumeric characters string. Obtain a GitHub personal access token here.

Examples

Good!

julia> sdad_setup!(db_user = "jbs3hp",
                   db_pwd = "MyVerySafePwd",
                   github_login = "Nosferican",
                   github_token = "0ipg0jvonteb54lv7j6cbgwn2snq3d3ac1pthxvz")
😃

Incomplete!

julia> sdad_setup!(db_user = "jbs3hp",
                   db_pwd = "MyVerySafePwd",
                   github_login = "Nosferican")
Warning: github_login has not been defined.
😞

Updating the configuration file.

julia> using GitHubAPI
julia> sdad_setup!(db_user = "jbs3hp",
                   db_pwd = "MyVerySafePwd",
                   github_login = "Nosferican",
                   github_token = "0ipg0jvonteb54lv7j6cbgwn2snq3d3ac1pthxvz")
julia> exit()
> julia
julia> using GitHubAPI # config up-to-date

Note

Updating the configuration file requires a restart of the session.

Updating the GitHub personal access token

Permanent through the configuration file

julia> sdad_setup!(github_token = "0ipg0jvonteb54lv7j6cbgwn2snq3d3ac1pthxvz")
😃

Temporarily

GitHubAPI.github_token = "0ipg0jvonteb54lv7j6cbgwn2snq3d3ac1pthxvz" # new value

Note

This method is temporary and will note overwrite the configuration file. Only the GitHub personal access token is allowed to be modified temporarily.

source

Identify some license which has not been scrapped yet.

The tables in the database related to monitoring the progress of the task are

universe.github_repos

universe.github_repos_tracker

GitHubAPI.get_licenses — Function.

get_licenses()::Vector{String}

List of SPDX ID for every OSI-approved license. (source: SPDX)

source

Run the scrapper

GitHubAPI.parse_repos! — Function.

parse_repos!(license::AbstractString,
             created_at::AbstractString = "2007-10-29T14:37:16+00:00",
             until::ZonedDateTime = until)
parse_repos!(license::AbstractString,
             data,
             as_of,
             created_at::AbstractString = "2007-10-29T14:37:16+00:00",
             until::ZonedDateTime = until)

For one a given license, it collects the information for each repository.

Arguments

license is a SPDX identifier for some OSI-approved license.

created_at is a datetime interval that GitHub supports (docs)

until: this value is the end-date for the queries (GitHubAPI.until)

Examples

Start querying for all the repositories with license zlib.

parse_repos!("zlib")

After checking the github_repos_tracker,

If the license scrapper already has completed some work, you can resume the job by passing the created_at argument.

parse_repos!("zlib",
             "2016-04-20T15:59:13+00:00..2017-12-17T09:59:36+00:00")

source

GitHubAPI.parse_commits! — Function.

parse_commits!(slug::AbstractString, until::AbstractString)
parse_commits!(slug::AbstractString, until::ZonedDateTime = until)

Writes to the database all the commit history for that repository.

Examples

Query one repository.

parse_commits!("uva-bi-sdad/GitHubAPI.jl")

For resuming work, find the date of the oldest commit in the database.

parse_commits!("uva-bi-sdad/GitHubAPI.jl", "2019-01-01 15:30:00+00")

source